+++ contracts/test/challengeV2/ChallengeEdgeLib.t.sol
@@ -9,7 +9,7 @@ import "./Utils.sol";
 import "../../src/challengeV2/libraries/ChallengeEdgeLib.sol";
 
 contract TestChallengeEdge {
-    function levelToType(uint256 level, uint256 numBigStepLevels) public pure returns (EdgeType eType) {
+    function levelToType(uint64 level, uint64 numBigStepLevels) public pure returns (EdgeType eType) {
         return ChallengeEdgeLib.levelToType(level, numBigStepLevels);
     }
 }
@@ -213,7 +213,7 @@ contract ChallengeEdgeLibTest is Test {
     }
 
     function testLevelToType() public {
-        uint256 numBigStep = 4;
+        uint64 numBigStep = 4;
         assertTrue(ChallengeEdgeLib.levelToType(0, numBigStep) == EdgeType.Block, "Block");
         assertTrue(ChallengeEdgeLib.levelToType(1, numBigStep) == EdgeType.BigStep, "Big step 1");
         assertTrue(ChallengeEdgeLib.levelToType(2, numBigStep) == EdgeType.BigStep, "Big step 2");
